+++ packages/base-entities/src/entities/base.entity.ts
@@ -8,6 +8,8 @@ import {
     UpdateDateColumn,
 } from 'typeorm';
 
+import { uuid } from '../types/uuid.type';
+
 /**
  * Base entity that all other entities extend
  */
@@ -15,19 +17,24 @@ export abstract class BaseEntity {
     /**
      * Id of the entity
      */
-    @PrimaryGeneratedColumn()
-    public id!: number;
+    @PrimaryGeneratedColumn({
+        type: 'uuid',
+    })
+    public id!: uuid;
 
     /**
      * When the entity was created
      */
-    @CreateDateColumn()
+    @CreateDateColumn({
+        name: 'created_at',
+    })
     public createdAt!: Date;
 
     /**
      * When the last time the entity was updated
      */
     @UpdateDateColumn({
+        name: 'updated_at',
         nullable: true,
     })
     public updatedAt?: Date;