+++ packages/base-entities/src/entities/index.ts
@@ -1,5 +1,0 @@
-/**
- * @copyright FLYACTS GmbH 2019
- */
-
-export { BaseEntity } from './base.entity';