+++ packages/job-runner/src/migrations/1587832806434-correct-created-at-and-updated-at-column-name-.sql
@@ -1,0 +1,4 @@
+ALTER TABLE public.jobs RENAME COLUMN "createdAt" TO created_at;
+ALTER TABLE public.jobs RENAME COLUMN "updatedAt" TO updated_at;
+ALTER TABLE public.job_queues RENAME COLUMN "createdAt" TO created_at;
+ALTER TABLE public.job_queues RENAME COLUMN "updatedAt" TO updated_at;