+++ packages/media-management/src/migrations/1587832806434-correct-created-at-and-updated-at-column-name-.sql
@@ -1,0 +1,4 @@
+ALTER TABLE public.files RENAME COLUMN "createdAt" TO created_at;
+ALTER TABLE public.files RENAME COLUMN "updatedAt" TO updated_at;
+ALTER TABLE public.media RENAME COLUMN "createdAt" TO created_at;
+ALTER TABLE public.media RENAME COLUMN "updatedAt" TO updated_at;